# API CRUD+Token Bearer dan Swagger/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
        "/api/mahasiswa": {
            "get": {
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Dapatkan daftar semua mahasiswa",
                "description": "Mengembalikan daftar semua mahasiswa",
                "operationId": "getMahasiswa",
                "responses": {
                    "200": {
                        "description": "Daftar mahasiswa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Mahasiswa"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Tambahkan mahasiswa baru",
                "description": "Menambahkan data mahasiswa baru",
                "operationId": "createMahasiswa",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Mahasiswa"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Mahasiswa berhasil ditambahkan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Mahasiswa"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Data tidak valid"
                    }
                }
            }
        },
        "/api/mahasiswa/{id}": {
            "put": {
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Perbarui data mahasiswa",
                "description": "Memperbarui data mahasiswa berdasarkan ID",
                "operationId": "updateMahasiswa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID mahasiswa yang akan diperbarui",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Mahasiswa"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mahasiswa berhasil diperbarui",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Mahasiswa"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Data tidak valid"
                    },
                    "404": {
                        "description": "Mahasiswa tidak ditemukan"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Hapus mahasiswa",
                "description": "Menghapus data mahasiswa berdasarkan ID",
                "operationId": "deleteMahasiswa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID mahasiswa yang akan dihapus",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mahasiswa berhasil dihapus"
                    },
                    "404": {
                        "description": "Mahasiswa tidak ditemukan"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Mahasiswa": {
                "title": "Mahasiswa",
                "description": "Model Mahasiswa",
                "properties": {
                    "id": {
                        "description": "ID mahasiswa",
                        "type": "integer"
                    },
                    "nama": {
                        "description": "Nama mahasiswa",
                        "type": "string"
                    },
                    "nim": {
                        "description": "Nomor Induk Mahasiswa",
                        "type": "string"
                    },
                    "email": {
                        "description": "Email mahasiswa",
                        "type": "string"
                    },
                    "jurusan": {
                        "description": "Jurusan mahasiswa",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Mahasiswa",
            "description": "Mahasiswa"
        }
    ]
}